Move exceptionHandlers to DefaultSlashCommandClient
